Use CachedState as param of trace creation
